refactor: remove adapter directory and moved NewDbClient to config package
